database/requirements.txt
# Python dependencies for EFRAG IG3 import
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
supabase==2.3.0
